Extract DelegateDrawable to be reusable in embedded.
